requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
is-safe-url==1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
python-dotenv==0.14.0
Werkzeug==1.0.1
